+++ ec2rlcore/prediag.py
@@ -65,6 +65,7 @@ def get_distro():
     distro = "unknown"
     alami_regex = re.compile(r"^Amazon Linux AMI release \d{4}\.\d{2}")
     alami2_regex = re.compile(r"^Amazon Linux (release \d \(Karoo\)|release \d.* \(\d{4}\.\d{2}\)|2)")
+    al2023_regex = re.compile(r"^Amazon Linux release 2023 \(Amazon Linux\)")
     rhel_regex = re.compile(r"^Red Hat Enterprise Linux*")
 
     # Amazon Linux & RHEL
@@ -76,6 +77,8 @@ def get_distro():
                 distro = ec2rlcore.constants.DISTRO_ALAMI
             elif re.match(alami2_regex, distro_str):
                 distro = ec2rlcore.constants.DISTRO_ALAMI2
+            elif re.match(al2023_regex, distro_str):
+                distro = ec2rlcore.constants.DISTRO_AL2023
             elif re.match(rhel_regex, distro_str) or \
                     re.match(r"^CentOS.*release (\d+)\.(\d+)", distro_str):
                 distro = ec2rlcore.constants.DISTRO_RHEL